refactor: divide test files as directory
